==> notebooks/00_notebook_template_ek.ipynb <==
## Notebook naming convention:

*XX_description_initials.ipynb*

Where XX = :
- **01**: Data gathering, initial processing, saving to raw data folder
- **02**: Data processing & EDA
- **03**: Individual model
- **04**: Comparing multiple models


# Import python packages
import os
import sys
import pandas as pd

# Add root & src directories to path
root_dir = os.path.join(os.getcwd(), '..')
src_dir = os.path.join(os.getcwd(), '..', 'src')
sys.path.append(root_dir)
sys.path.append(src_dir)

# Import local secrets & functions
from conf.auth_example import api_key
from d00_utils.example import print_hello

print(api_key)

print_hello()

# Load data
df = pd.read_csv('../data/example_current_df.csv', index_col=0)
df
