==> notebooks/01_frankenvariable_iu.ipynb <==
Unfinished - changed approach. See - firstpassb notebook.

# Reading in data

import os
import pandas as pd
import numpy as np
from datetime import datetime
import seaborn as sns
import re 
from pandas_profiling import ProfileReport

import matplotlib.pyplot as plt

path = 'alldata/'


def make_dataframes(path):
    file_names = [pos_csv for pos_csv in os.listdir(path) if pos_csv.endswith('.csv')]
    dataframes = {}
    dfs = []
    for i, file in enumerate(file_names):
        df_name = ("df_" + file)[:-4]
        dfs.append(pd.read_csv(path + file))
        dataframes[df_name] = dfs[i]
    return dataframes, file_names

data_dict, files = make_dataframes(path)

# We created a dictionary of dataframes:
data_dict.keys()
# These are the dataframes:

# Pre-merge EDA

# We join them on the time_bin column:
data_dict['df_image_label_isa'] = data_dict['df_image_label_isa'].rename(columns={"timetaken": "time_bin"})

def fix_time_format(df):
    """ 
    Turn format to same format for df with 'time_bin' datetime object column.
    """
    print("before:", df['time_bin'].dtype)
    df['time_bin']= pd.to_datetime(df['time_bin']) 
    print("after:", df['time_bin'].dtype)
#    df['time_bin'] = df['time_bin'].dt.strftime('%Y-%m-%d %H:%M')


# Read in pkl tweets into dataframe:
path_to_pkl = "alldata/geotweets_labeled_binned.pkl"

def read_pkl(path_to_pkl):
    tweets = pd.read_pickle(path_to_pkl)
    
    # Create tweets_safe (time bin column, safe score column)
    print(tweets.dtypes)
   # tweets['time_bin'] = tweets['time_bin'].astype(str)
    tweets_safe = tweets[pd.notnull(tweets['safe'])]
#     tweets_safe = tweets.groupby('time_bin').agg(avg_safe=('safe', 'mean'), num_tweets =('time_bin', 'count'))
#     #tweets_safe = tweets.groupby('time_bin').agg({'safe':['mean'], 'safe':['sum']})
#     print(len(tweets_safe))
#     tweets_safe = tweets_safe.drop(columns=['num_tweets'])
#     tweets_safe.columns
    tweets_safe = tweets_safe.reset_index()
    
    # tweets need to be changed to stardard labels: safe = 0, unsafe = 1
    tweets_safe["safe"] = tweets_safe["safe"].replace([0, 1], [1,0])
    
    #fix_time_format(tweets_safe)

    
    # create tweets_count (time bin column, number of tweets)
    tweets_count = tweets.groupby('time_bin').agg(avg_safe=('safe', 'mean'), num_tweets =('time_bin', 'count'))
    tweets_count = tweets_count.drop(columns=['avg_safe'])
    tweets_count = tweets_count.reset_index()
    
    
    return tweets_safe, tweets_count

tweets_safe, tweets_count = read_pkl(path_to_pkl)

considered_dfs = ['df_curpop_df',
                  'df_binned_current_log',
                  'df_image_label_isa',
                  'df_Survey_clean',
                  'df_binned_weather_history']

def convert_and_examine_time(considered_dfs):
    
    for df in considered_dfs:
        print("df name:", df)
        fix_time_format(data_dict[df])
        print('begins:', data_dict[df]['time_bin'].min(), 
              "\nends:", data_dict[df]['time_bin'].max(), 
              "\nnon-na vals:", data_dict[df]['time_bin'].count(), "\n\n" )
    
    print("df name:", "tweets_safe")
    print('begins:', tweets_safe['time_bin'].min(), 
          "\nends:", tweets_safe['time_bin'].max(), 
          "\nnon-na vals:", tweets_safe['time_bin'].count())

# fix_time_format(data_dict['df_curpop_df'])
# fix_time_format(data_dict['df_binned_current_log'])
# fix_time_format(data_dict['df_image_label_isa'])
# fix_time_format(data_dict['df_Survey_clean'])
# fix_time_format(data_dict['df_binned_weather_history'])

convert_and_examine_time(considered_dfs)

# Make one big dataframe

# Join dataframes on time_bin column:

df_full = pd.merge(data_dict['df_curpop_df'], data_dict['df_binned_weather_history'], on='time_bin', how='outer')
df_full = pd.merge(df_full, data_dict['df_image_label_isa'], on='time_bin', how='outer')
df_full = pd.merge(df_full, data_dict['df_image_label_isa'], on='time_bin', how='outer')
df_full = pd.merge(df_full, data_dict['df_binned_current_log'], on='time_bin', how='outer')

df_full = pd.merge(df_full, tweets_safe, on='time_bin', how='outer')
df_full = pd.merge(df_full, tweets_count, on='time_bin', how='outer')


df_full.head(1)

# Post-merge EDA

# Pandas profiling Report
profile = ProfileReport(df_full, title='report', html={'style':{'full_width':True}})
profile.to_file(output_file="report.html")

# We can see we are missing a lot of data for some variables:
df_full.count()

df_full.shape

Project criteria: **We will only include data from March 23rd, 2020 onward**:

df = df_full[df_full['time_bin'] >= '3/23/2020 00:00']
print("before filtering:", df_full.shape, "\nafter filtering:", df.shape)

# Building a target variable, and some feature engineering

We have a few variables we can use to build a labelled column - we have a lot of missing values, so we can cobble a target variable using different columns:

- main_track: from survey data on how busy the main track of the park is (df_Survey_clean)
- label, Isa: from photos (df_image_label_isa)
- tweets_safe: from twitter scraping data


target_var = ['main_track', 'label', 'Isa', 'tweets_safe']
indep_vars = ['activity', 'current_popularity']

print(df['time_bin'].nunique())
print(len(df))

Naively drop any target variables that we don't have zeros or ones for:

data_dict.keys()
temperature, precipitation, windspeed, detailed_status, current_popularity

data_dict['df_binned_weather_history']
