==> notebooks/01_OWM_api_ek.ipynb <==
import os
import sys
import pyowm
import pandas as pd
import numpy as np
import json
import requests
from pyowm.owm import OWM
from pyowm.weatherapi25.forecast import Forecast

root_dir = os.path.join(os.getcwd(), '..')
src_dir = os.path.join(os.getcwd(), '..', 'src')
sys.path.append(root_dir)
sys.path.append(src_dir)

from conf.auth import OWM_api_key

owm = pyowm.OWM(OWM_api_key)

park_loc = pd.DataFrame({
    'name': ['Prospect Park', 'Wooster Square'],
    'lat': [40.662551, 41.304745],
    'long': [-73.969256, -72.917757]})

park_name = 'Prospect Park'
park_lat = park_loc.loc[park_loc.name == park_name, 'lat'].iloc[0]
park_long = park_loc.loc[park_loc.name == park_name, 'long'].iloc[0]

mgr = owm.weather_manager()
park_forecast = mgr.forecast_at_coords(lat=float(park_lat), lon=float(park_long), interval='3h').forecast
park_forecast

 park_forecast.reception_time()

park_forecast.get(1).to_dict()

OWM_forecast_cols = [
    'park_name', 'reception_time', 'reference_time', 'sunset_time',
    'sunrise_time', 'clouds', 'rain', 'snow', 'wind_speed',
    'wind_deg', 'humidity', 'press', 'press_sea', 'temp',
    'temp_feels', 'temp_max', 'temp_min', 'status', 'detailed_status']

def forecast_to_df(park_forecast, t_sunrise, t_sunset, tz_correct):
    ''' Takes a forecast object and sunrise/sunset time,
        Returns datafram of relevant info
    '''
    # Neet to adjust sunrise/set times, should only go
    # into columns of the same day (or just use last day in window?)
    
    forecast_df = pd.DataFrame(columns=OWM_forecast_cols)

    for x in park_forecast:
        if (t_sunrise == 0) or (t_sunset == 0):
            x_sunrise = np.nan
            x_sunset = np.nan
        else:
            x_sunrise = t_sunrise + tz_correct
            x_sunset = t_sunset + tz_correct

        x_rain = 0
        if x.rain:
            x_rain = x.rain['3h']

        x_snow = 0
        if x.snow:
            x_snow = x.snow['3h']

        to_add_dict = {
            'park_name': [park_name],
            'reception_time': [park_forecast.reception_time('unix') + tz_correct],
            'reference_time': [x.reference_time() + tz_correct],
            'sunrise_time': [x_sunrise],
            'sunset_time': [x_sunset],
            'clouds': [x.clouds],
            'rain': [x_rain],
            'snow': [x_snow],
            'wind_speed': [x.wind(unit='meters_sec')['speed']],
            'wind_deg': [x.wind(unit='meters_sec')['deg']],
            'humidity': [x.humidity],
            'press': [x.pressure['press']],
            'press_sea': [x.pressure['sea_level']],
            'temp': [x.temperature(unit='fahrenheit')['temp']],
            'temp_feels': [x.temperature(unit='fahrenheit')['feels_like']],
            'temp_max': [x.temperature(unit='fahrenheit')['temp_max']],
            'temp_min': [x.temperature(unit='fahrenheit')['temp_min']],
            'status': [x.status],
            'detailed_status': [x.detailed_status]}
        
        to_add = pd.DataFrame(to_add_dict)
        forecast_df = forecast_df.append(to_add, ignore_index=True)
    
    return forecast_df

forcast_df = forecast_to_df(park_forecast, 0, 0, 10)
forcast_df.head()

Errors in pyowm means I can't get sunrise/sunset time directly, so moving to:

- Calling the api manually for the forcast JSON
- Extracting the sunrise/sunset time
- Passing the JSON back into the pyowm parser then use the work from above

api_url = 'http://api.openweathermap.org/data/2.5/forecast'
payload = {
    'lat': park_lat,
    'lon': park_long,
    'appid': OWM_api_key}
r = requests.get(api_url, params=payload)

d = json.loads(r.text)

sunrise_time = d['city']['sunrise']
sunset_time = d['city']['sunset']
tz_correct = d['city']['timezone']

# weather_df = forecast_to_df(forecastparser.ForecastParser().parse_JSON(r.text), sunrise_time, sunset_time, tz_correct)
weather_df = forecast_to_df(Forecast.from_dict(json.loads(r.text)), sunrise_time, sunset_time, tz_correct)

weather_df.head(5)

The dataframe weather_df is what I'll be adding to the log of forcast data.

weather_df.to_csv('example_weather_df.csv')
