==> notebooks/02_build_dataframe_for_modeling_ee.ipynb <==
# Integrate Our Separate Data Streams into One Data Frame

import pandas as pd
import numpy as np
import datetime as dt

intro = "../data/"
tweets = pd.read_pickle(intro+"02_geotweets_labeled_ee.pkl")
curpop = pd.read_csv(intro+"curpop_df.csv", index_col = 0)
weekpop = pd.read_csv(intro+"weekpop_df.csv", index_col = 0)
img = pd.read_csv(intro+"image_label.csv", index_col = 0)
curlog = pd.read_csv(intro+"current_log.csv")
wlog = pd.read_csv(intro+"binned_weather_history.csv")
survey = pd.read_csv(intro+"Survey_clean.csv")

### Obtain a Data Frame of Time Bins Labeled by Tweet Labels

# tweets.head()

# Note: in some cases, we have multiple tweets for a single time bin.
# In what follows, it'll make life easier if we swap -1 for 2 as an indicator of 
# no information. We will later swap 2 back for -1 to match everyone else. 
safe = []
for i in range(0,len(tweets)):
    value = tweets.iloc[i].safe
    if value == 2:
        safe.append(-1)
    else:
        safe.append(value)

tweets['safe'] = safe

bin_counts = tweets.groupby('time_bin')['safe'].agg('count')
bin_counts
bigbins = bin_counts[bin_counts > 1]
bigbins = bigbins.reset_index()
bigbin_tweets = pd.merge(left = bigbins.time_bin, right = tweets, left_on = ['time_bin'], right_on = ['time_bin'])

# Now we can check manually whether we have any bins containing tweets with polar opposite labels
# as opposed to either 0 or 1 + uninformative.
# Note: this is a substantive decision, since it means that if some tweets are informative but others not, we 
# label the tweet according to the information we got from the informative tweets, rather than ignoring this 
# information. 

for i in range(0,len(bigbin_tweets)-1):
    if bigbin_tweets.iloc[i].safe + bigbin_tweets.iloc[i+1].safe == 1:
        print(i, ":", bigbin_tweets.iloc[i], i+1, ":", bigbin_tweets.iloc[i+1])

# Again, here we are checking to see whether we have any adjacent rows with opposite labels: 
# 0 next to 1 or vice versa.
# There are none. So we can safely assign each bin the max of the safe values that occur in that bin.
tweet_bin_labels = tweets.groupby('time_bin')['safe'].agg('max')
# This is why we changed NaN to -1. 

tweet_bin_labels = tweet_bin_labels.reset_index()
# tweet_bin_labels

# Standardize interpretation of labels: 0 = safe, 1 = unsafe, 2 = unsure
# Add column indicating if it's a weekend (1) or not (0)

safe = []
weekend = []

for i in range(0,len(tweet_bin_labels)):
    value = tweet_bin_labels.iloc[i].safe
    if value == -1.0:
        safe.append(int(2))
    else:
        safe.append(value)

tweet_bin_labels['safe'] = safe

# Check that it all worked
# tweet_bin_labels.head(10)

### Preprocess the Current Popularity Data

# curpop.head()

# Convert the time bin to datetime
curpop['binned'] = pd.to_datetime(curpop['binned'])
curpop = curpop.reset_index()

curpop = curpop[['current_popularity','binned']]
curpop = curpop.groupby('binned').agg('mean')
curpop = curpop.reset_index()
# curpop

### Preprocess the Average Time Popularity data

# weekpop.head()

# lengths = []
# for i in range(0,5):
#    lengths.append(len(weekpop.iloc[i]['data']))
# lengths

## Ask Huayi about this! Why do the 'data' entries of weekpop have different lengths?

### Preprocess the Image Labels

# img.head()

img['timetaken'] = pd.to_datetime(img['timetaken'])
img.columns = ['timetaken','img_label']
# img.head()

### Preprocess the Weather Data

#curlog.columns

#wlog.columns

# Convert all time columns to datetime
curlog['sunrise_time'] = pd.to_datetime(curlog['sunrise_time'])
curlog['sunset_time'] = pd.to_datetime(curlog['sunset_time'])
curlog['time_bin'] = pd.to_datetime(curlog['time_bin'])
wlog['time_bin'] = pd.to_datetime(wlog['time_bin'])

# Restrict wlog to the dates since the crisis and drop the unnecessary 'time' column
wlog = wlog.set_index('time_bin')
wlog = wlog.sort_index()
wlog = wlog["2020-03-23":]
wlog = wlog.reset_index()
wlog = wlog.drop(['time'], axis = 1)

# To avoid conflicts, drop the 'humidity' column of curlog
curlog = curlog.drop('humidity',axis=1)

#curlog.head()

#wlog.head()

### Preprocess the Survey Data

#survey.head()

survey['time_bin'] = pd.to_datetime(survey['time_bin'])

# len(survey)

### Join our dataframes into one big dataframe

# Index all dataframes by their time bins for easy joining

tweet_bin_labels.set_index('time_bin', inplace = True)
curpop.set_index('binned', inplace = True)
img.set_index('timetaken', inplace = True)
curlog.set_index('time_bin', inplace = True)
wlog.set_index('time_bin', inplace = True)
survey.set_index('time_bin', inplace = True)

df = tweet_bin_labels.join(curpop, how = 'outer')\
.join(img, how = 'outer')\
.join(curlog, how = 'outer')\
.join(wlog, how = 'outer')\
.join(survey, how = 'outer')

# df.head()

# How often is it that 'safe' is null and the survey responses informative?
# df[(pd.notnull(df.safe)) & (pd.notnull(df.main_track)) & (pd.notnull(df.not_main_track))]

# # How often is it unsafe on the main track but safe off of it?
# df[(pd.notnull(df.main_track)) & (pd.notnull(df.not_main_track)) & (df.main_track != df.not_main_track) & (df.main_track != 2.0) & (df.not_main_track != 2.0)]

# Use image label and survey data to improve the 'safe' column
safe = []

# First, let substantive survey data, where it exists, overrule tweets data
for i in range(0, len(df)):
    
    # For timebin i, let:
    #  s be the current safety label
    #  m be the main track's safety
    #  nm be off-main-track safety
    #  im be the image label if it exists
    
    s = df.iloc[i].safe
    m = df.iloc[i].main_track
    nm = df.iloc[i].not_main_track
    im = df.iloc[i].img_label
    
        
    ## Make the survey labels dominate the tweet labels:
        
    # If both m and nm are informative (neither NaN nor 2.0), then:
    if (pd.notnull(m) & pd.notnull(nm) & (m != 2.0) & (nm != 2.0)):
        # append the larger of the two (so: unsafe dominates safe; err on side of caution)
        safe.append(max(m,nm))
        
    # If m is informative but nm is not, append m:
    elif (pd.notnull(m) & (m != 2.0)):
        safe.append(m)
        
    # If nm is informative but m is not, append nm:
    elif (pd.notnull(nm) & (nm != 2.0)):
        safe.append(nm)


    ## If survey values are both null, use image labels. These don't dominate tweets.
        
    # If m and nm are both null, and safe value is 2 (so marked uninformative), and 
    # image label is not null, append image label:
    elif ((s == 2.0) & pd.notnull(im)):
        safe.append(im)
        
    # If m and nm are both null and either s is informative or im is uninformative, append s
    else: 
        safe.append(s)

# Now, replace all NaNs with 2.0 
for i in range(0,len(safe)):
    if pd.isnull(safe[i]):
        safe[i] = 2.0

df['safe'] = safe

#df.head()

#pd.unique(curlog.detailed_status)

# Drop unnecessary columns
df = df.drop(['live','activity','status','main_track','not_main_track', 'park_name', 'img_label'],axis=1)

# One-hot encode detailed weather statuses
df = pd.get_dummies(data = df, prefix='', prefix_sep='',columns = ['detailed_status'],dummy_na = True)
# Note: This fills NaNs with 0s-- we will need to fix that

# Where any of the weather columns should be NaN, replace with NaN
statuses = {'light rain':[], 'overcast clouds':[], 'few clouds':[], 'broken clouds':[],
           'scattered clouds':[], 'clear sky':[], 'moderate rain':[], 'mist':[],
           'thunderstorm with rain':[]}

for i in range(0,len(df)):
    if df.iloc[i]['nan'] == 1:
        for status in statuses.keys():
            statuses[status].append(np.NaN)
    else:
        for status in statuses.keys():
            statuses[status].append(df.iloc[i][status])

for status in statuses.keys():
    df[status] = statuses[status]
    
# Drop the now-unnecessary nan column
df = df.drop('nan',axis = 1)

#df.head()

#df.info()

# i = 1
# for column in df.columns:
#     if len(pd.unique(df[column])) > 3:
#         print(i, ":", column, ":", len(pd.unique(df[column])))
#         i += 1

import matplotlib.pyplot as plt
plt.style.use("ggplot")

fig = plt.figure(figsize=(10,10))

for i in range(0,len(df.columns)):
    if len(pd.unique(df[df.columns[i]])) > 3:
        plt.subplot(5,5,i+1)
        plt.hist(df[df.columns[i]])
        plt.xticks(rotation = 45)
        plt.title(df.columns[i])
#         plt.ylabel("# timebins with that value")
#         plt.xlabel("value for" + column)
plt.tight_layout()
plt.savefig("../src/d06_visuals/predictor_distributions_ee.png")
#plt.show()

# Apply a log transform to our exponentially decaying variables

def safe_ln(x):
    if x <= 0:
        return 0
    return np.log(x)

df['rain_1h'] = df['rain_1h'].apply(lambda x: safe_ln(x))
df['precipitation'] = df['precipitation'].apply(lambda x: safe_ln(x))

df.to_pickle(intro+"03_final_multipurpose_df_ee.pkl")
