==> notebooks/02_EDA2_iu.ipynb <==
## Setup

import pandas as pd
import os

###### DEBUGGING TOOLS:

# Change debug_bool to True/False to turn on/off debug prints
debug_bool = True
def debug_print(*args):
    if debug_bool:
        print('debug print:', '\n', *args, '\n')

## Import csv files



path_to_data = '../data'



def create_filelist(path_to_data):
    file_names = [pos for pos in os.listdir(path_to_data) if pos.endswith('.csv')]
    debug_print('len of file_names:', len(file_names))
    return file_names

def build_df(file_names, path_to_data):
    for file in file_names
    pd.read_csv("filename.csv")
