==> notebooks/02_EDA_iu.ipynb <==
## Setup

import pandas as pd
import os
import glob

###### DEBUGGING TOOLS:

# Change debug_bool to True/False to turn on/off debug prints
debug_bool = True
def debug_print(*args):
    if debug_bool:
        print('debug print:', '\n', *args, '\n')

## Import csv files

Folder 'data' has csv files with tweets from tweepy.

df = pd.concat(map(pd.read_csv, glob.glob('data/*.csv')))

print(df.shape)
df.head()

## Searching for tweets:

import re
distancing = []
d_loc = []

for i in range(0,len(df)):
    if re.match('.*(distancing)|(social)|(distance)|(crowd)|(crowds)|(crowded)|(busy)|(full)|(too many)|(many)|(packed)|(close)|(groups)|(safe)|(empty)|(quarantine).*',df.iloc[i].Tweet):
        distancing.append(1)
        d_loc.append(i)
        
        
print("Found",len(distancing),"tweets related to the crowding level in Prospect Park: \n")
for i in range(len(d_loc)):
    print('Tweet at time', df.iloc[d_loc[i]].Date, ":", df.iloc[d_loc[i]].Tweet, "\n")


Tweets about prospect park:

distancing = []
d_loc = []

for i in range(0,len(df)):
    if re.match('.*(Prospect Park).*',df.iloc[i].Tweet):
        distancing.append(1)
        d_loc.append(i)
        
        
print("Found",len(distancing),"tweets related to the crowding level in Prospect Park: \n")
for i in range(len(d_loc)):
    print('Tweet at time', df.iloc[d_loc[i]].Date, ":", df.iloc[d_loc[i]].Tweet, "\n")
