==> notebooks/02_geotweets_binner_ee.ipynb <==
### Use Huayi's time binning function to assign our geotagged tweets to 15-minute time bins:
#### 05, 20, 35, and 50 mins past the hour. 

import pandas as pd
import numpy as np
import datetime as dt

# Huayi's time binning function! 

from datetime import timedelta 

def binMinute(minute):
    '''
    Calculate which of the 15min bins the current minute falls in.
    The bins are 05, 20, 35, 50. 
    Current minutes less than 05 belong to bin 50 of the hour before. 
    '''    
    m2q= (minute-5)/15
    if m2q < 0:
        newminute = -10 
    elif m2q < 1:
        newminute = 5
    elif m2q < 2:
        newminute = 20
    elif m2q < 3:
        newminute = 35
    elif m2q < 4:
        newminute = 50
    return newminute       
    
def BinTime(df_datetime_col):
    '''
    Take a df column that contains datetime object (df_datetime_col)
    Return a new column that contains the binned datetime object
    The bins are as defined in binMinute
    '''
    binned_col= df_datetime_col.apply(lambda x: x.replace(minute = 0) + timedelta(minutes=binMinute(x.minute)))
    return binned_col

df = pd.read_pickle("../data/02_geotweets_labeled_ee.pkl")

# df.head()

df['time_bin'] = BinTime(df.date.dt.floor('min'))

# df.head()

df = df.reset_index(drop=True)
# df.head()

# Localize time bin (remove comparison to UTC)

time_bin = []
for i in range(0, len(df)):
    bn = df.iloc[i].time_bin.tz_convert(None) + pd.Timedelta(-4,'h')
    time_bin.append(bn)

df['time_bin'] = time_bin

df.head()

len(df)

df.to_pickle("../data/02_geotweets_labeled_binned_ee.pkl")
