==> notebooks/02_geotweets_EDA_2_ee.ipynb <==
import pandas as pd
import numpy as np

intro = "../data/"
df = pd.read_pickle(intro+"03_final_multipurpose_df_ee.pkl")

# print("# safe:", len(df[df.safe == 0]))
# print("# unsafe:", len(df[df.safe == 1]))
# print("# unknown:", len(df[df.safe == 2]))

# df.info()

# for column in df.columns:
#     print(column, ":", len(pd.unique(df[column])))

import matplotlib.pyplot as plt
plt.style.use("ggplot")
for column in df.columns:
    if len(pd.unique(df[column])) > 3:
        plt.hist(df[column])
        plt.title(column)
        plt.ylabel("# timebins with that value")
        plt.xlabel("value for " + column)
        plt.savefig("../visualizations/" + column + ".png")
        #plt.show()
