==> notebooks/02_geotweets_labeler_ee.ipynb <==
### Label the tweets as unsafe for visiting the park (0), safe (1), or NaN
### NOTE: This labeling convention will be reversed at the end

import pandas as pd
import numpy as np
import datetime as dt

df = pd.read_pickle("../data/01_geotweets_raw_since_crisis_began_ee.pkl")

len(df)

### Initiate an empty dict which we will fill with row labels, in groups of 10.
rows = {}

for i in range(0,10):
    print(i, ' : ', df.iloc[i].text)

# None of these rows is informative! They're all garbage. 
rows[0] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[0])

for i in range(10,20):
    print(i, ' : ', df.iloc[i].text)

# Row #13 shows a picture with lots of people on a main path! Not safe. 
rows[1] = [np.NaN,np.NaN,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[1])

for i in range(20,30):
    print(i, ' : ', df.iloc[i].text)

rows[2] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[2])

for i in range(30,40):
    print(i, ' : ', df.iloc[i].text)

rows[3] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[3])

for i in range(40,50):
    print(i, ' : ', df.iloc[i].text)

rows[4] = [np.NaN,np.NaN,0,np.NaN,1,1,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[4])

for i in range(50,60):
    print(i, ' : ', df.iloc[i].text)

rows[5] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,0,np.NaN,np.NaN,1]

len(rows[5])

for i in range(60,70):
    print(i, ' : ', df.iloc[i].text)

rows[6] = [np.NaN,np.NaN,1,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[6])

for i in range(70,80):
    print(i, ' : ', df.iloc[i].text)

rows[7] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[7])

for i in range(80,90):
    print(i, ' : ', df.iloc[i].text)

rows[8] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[8])

for i in range(90,100):
    print(i, ' : ', df.iloc[i].text)

rows[9] = [np.NaN,1,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,0]

len(rows[9])

for i in range(100,110):
    print(i, ' : ', df.iloc[i].text)

rows[10] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,1,np.NaN,1,0]

len(rows[10])

for i in range(110,120):
    print(i, ' : ', df.iloc[i].text)

rows[11] = [np.NaN,np.NaN,np.NaN,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[11])

for i in range(120,130):
    print(i, ' : ', df.iloc[i].text)

rows[12] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[12])

for i in range(130,140):
    print(i, ' : ', df.iloc[i].text)

rows[13] = [np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[13])

for i in range(140,150):
    print(i, ' : ', df.iloc[i].text)

rows[14] = [np.NaN,np.NaN,np.NaN,np.NaN,0,np.NaN,np.NaN,1,np.NaN,1]

len(rows[14])

for i in range(150,160):
    print(i, ' : ', df.iloc[i].text)

rows[15] = [np.NaN,np.NaN,0,np.NaN,np.NaN,1,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[15])

for i in range(160,170):
    print(i, ' : ', df.iloc[i].text)

rows[16] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[16])

for i in range(170,180):
    print(i, ' : ', df.iloc[i].text)

rows[17] = [np.NaN,np.NaN,0,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[17])

for i in range(180,190):
    print(i, ' : ', df.iloc[i].text)

rows[18] = [np.NaN,np.NaN,np.NaN,1,0,0,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[18])

for i in range(190,200):
    print(i, ' : ', df.iloc[i].text)

rows[19] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[19])

for i in range(200,210):
    print(i, ' : ', df.iloc[i].text) 

rows[20] = [np.NaN,np.NaN,1,np.NaN,1,np.NaN,np.NaN,0,np.NaN,np.NaN]

len(rows[20])

for i in range(210, 220):
    print(i, ' : ', df.iloc[i].text)

rows[21] = [np.NaN,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[21])

for i in range(220, 230):
    print(i, ' : ', df.iloc[i].text)

rows[22] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[22])

for i in range(230, 240):
    print(i, ' : ', df.iloc[i].text)

rows[23] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,0,np.NaN,np.NaN]

len(rows[23])

for i in range(240, 250):
    print(i, ' : ', df.iloc[i].text)

rows[24] = [np.NaN,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[24])

for i in range(250, 260):
    print(i, ' : ', df.iloc[i].text)

rows[25] = [np.NaN,np.NaN,1,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[25])

for i in range(260, 270):
    print(i, ' : ', df.iloc[i].text)

rows[26] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[26])

for i in range(270, 280):
    print(i, ' : ', df.iloc[i].text)

rows[27] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[27])

for i in range(280, 290):
    print(i, ' : ', df.iloc[i].text)

rows[28] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[28])

for i in range(290, 300):
    print(i, ' : ', df.iloc[i].text)

rows[29] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,1,np.NaN]

len(rows[29])

for i in range(300, 310):
    print(i, ' : ', df.iloc[i].text)

rows[30] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,1,np.NaN]

len(rows[30])

for i in range(310, 320):
    print(i, ' : ', df.iloc[i].text)

rows[31] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[31])

for i in range(320, 330):
    print(i, ' : ', df.iloc[i].text)

rows[32] = [np.NaN,0,np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[32])

for i in range(330, 340):
    print(i, ' : ', df.iloc[i].text)

rows[33] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[33])

for i in range(340, 350):
    print(i, ' : ', df.iloc[i].text)

rows[34] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[34])

for i in range(350, 360):
    print(i, ' : ', df.iloc[i].text)

rows[35] = [np.NaN,0,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[35])

for i in range(360, 370):
    print(i, ' : ', df.iloc[i].text)

rows[36] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[36])

for i in range(370, 380):
    print(i, ' : ', df.iloc[i].text)

rows[37] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,1,np.NaN]

len(rows[37])

for i in range(380, 390):
    print(i, ' : ', df.iloc[i].text)

rows[38] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[38])

for i in range(390, 400):
    print(i, ' : ', df.iloc[i].text)

rows[39] = [np.NaN,np.NaN,np.NaN,np.NaN,1,1,1,np.NaN,np.NaN,np.NaN]

len(rows[39])

for i in range(400, 410):
    print(i, ' : ', df.iloc[i].text)

rows[40] = [np.NaN,np.NaN,np.NaN,np.NaN,np.NaN,0,1,np.NaN,np.NaN,np.NaN]

len(rows[40])

for i in range(410, len(df)):
    print(i, ' : ', df.iloc[i].text)

rows[41] = [np.NaN,np.NaN,np.NaN,np.NaN]

len(rows[41])

len(df)

### Now we have finished labeling all our rows. 
### Gather the row labels into a single long column of row labels

# Initialize our response variable
safe = []

# Concatenate all our rows into one long column
for i in range(0,max(rows.keys())+1):
    safe = safe + rows[i]

len(safe)

df['safe'] = safe

### Filter out the uninformative garbage / bot tweets that skew the tweet time aggregations, and add the row that Isabel found.

# Filter out the garbage tweets by dannyhellman
df = df[df.username != 'dannyhellman']

import datetime as dt
import pytz
est = pytz.timezone('US/Eastern')
pd.to_datetime("2020-04-04 16:26:40").tz_localize(est)

# Add the tweet that Isabel found
import datetime as dt
import pytz
est = pytz.timezone('US/Eastern')
date = pd.to_datetime("2020-04-04 16:26:40").tz_localize(est)

row = pd.Series({"date":date,
                 "text":"Prospect Park is packed. Unreal.",
                 "username":"",
                 "day_of_week":"Saturday",
                 "hour":int(16),
                 "safe":int(0)})

row_df = pd.DataFrame([row], index = [date])

df = pd.concat([row_df, df])
df = df.sort_index()           

# Check that it worked
df.iloc[111:116]

len(df)

### Standardize interpretation of labels to match everyone else's: 
#### 0 = safe, 1 = unsafe, 2 = unsure

safe = []

for i in range(0,len(df)):
    value = df.iloc[i].safe
    if pd.isnull(value):
        safe.append(int(2))        
    else:
        safe.append(int(1 - value))

df['safe'] = safe

df.iloc[111:116]

### Save our dataframe

# Save our labels
df.to_pickle("../data/02_geotweets_labeled_ee.pkl")
